json file code updated

+++ webScrapping/espn_scrapper/scorecards.js
@@ -128,12 +128,17 @@ function getMatchDetails(html){
                     oppTeam=selecTool(teamNames[0]).text();
                 }
                 let ownTeamOppTeam=ownTeam+" VS "+oppTeam;
+
                 processInformation(ownTeam,oppTeam,venue,date,result,batsman,totalRuns,totalBalls,total4s,total6s,sr);
 
+
+
                 let obj={ownTeamOppTeam,venue,date,result,batsman,totalRuns,totalBalls,total4s,total6s,sr};
-                
+            
                 const data = JSON.stringify(obj);
-                fs.writeFileSync("scorecards.json",data,{flag:'a'});
+                if(!fs.existsSync("scorecards.json")){
+                    fs.writeFileSync("scorecards.json",data,{flag:'a'});
+                }
                 // fs.writeFileSync("scorecard.txt",data,{flag:'a'});
                 // fs.writeFile("scorecard.txt",data,{flag:'a'},cb);
                 // function cb(err,data){
